feat(eslint-plugin-template): export template parser services (#888)
