+++ packages/eslint-plugin-template/src/rules/accessibility-alt-text.ts
@@ -2,10 +2,8 @@ import type {
   Node,
   TmplAstElement,
 } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { getTemplateParserServices } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 import { getAttributeValue } from '../utils/get-attribute-value';
 
 type Options = [];
