+++ packages/eslint-plugin-template/src/rules/accessibility-label-has-associated-control.ts
@@ -1,8 +1,6 @@
 import type { TmplAstElement } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { getTemplateParserServices } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 import { isChildNodeOf } from '../utils/is-child-node-of';
 
 type LabelComponent = {
