+++ packages/eslint-plugin-template/src/rules/accessibility-role-has-required-aria.ts
@@ -2,12 +2,10 @@ import type {
   TmplAstTextAttribute,
   TmplAstElement,
 } from '@angular-eslint/bundled-angular-compiler';
+import { getTemplateParserServices } from '@angular-eslint/utils';
 import type { ARIARoleDefinitionKey } from 'aria-query';
 import { roles } from 'aria-query';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { createESLintRule } from '../utils/create-eslint-rule';
 import { getDomElements } from '../utils/get-dom-elements';
 import { toPattern } from '../utils/to-pattern';
 import { isSemanticRoleElement } from '../utils/is-semantic-role-element';
