+++ packages/eslint-plugin-template/src/rules/accessibility-valid-aria.ts
@@ -7,12 +7,10 @@ import {
   TmplAstBoundAttribute,
   TmplAstTextAttribute,
 } from '@angular-eslint/bundled-angular-compiler';
+import { getTemplateParserServices } from '@angular-eslint/utils';
 import type { ARIAProperty, ARIAPropertyDefinition } from 'aria-query';
 import { aria } from 'aria-query';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { createESLintRule } from '../utils/create-eslint-rule';
 import { getDomElements } from '../utils/get-dom-elements';
 import { toPattern } from '../utils/to-pattern';
