+++ packages/eslint-plugin-template/src/rules/banana-in-box.ts
@@ -1,8 +1,6 @@
 import type { TmplAstBoundEvent } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { getTemplateParserServices } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 type Options = [];
 export type MessageIds = 'bananaInBox';
