+++ packages/eslint-plugin-template/src/rules/cyclomatic-complexity.ts
@@ -2,10 +2,8 @@ import type {
   TmplAstBoundAttribute,
   TmplAstTextAttribute,
 } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { getTemplateParserServices } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 type Options = [{ maxComplexity: number }];
 export type MessageIds = 'cyclomaticComplexity';
