+++ packages/eslint-plugin-template/src/rules/no-any.ts
@@ -4,10 +4,8 @@ import {
   PropertyRead,
   ThisReceiver,
 } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  ensureTemplateParser,
-} from '../utils/create-eslint-rule';
+import { ensureTemplateParser } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 
 type Options = [];
 export type MessageIds = 'noAny' | 'suggestRemoveAny';
