+++ packages/eslint-plugin-template/src/rules/no-positive-tabindex.ts
@@ -3,10 +3,8 @@ import type {
   TmplAstBoundAttribute,
   TmplAstTextAttribute,
 } from '@angular-eslint/bundled-angular-compiler';
-import {
-  createESLintRule,
-  getTemplateParserServices,
-} from '../utils/create-eslint-rule';
+import { getTemplateParserServices } from '@angular-eslint/utils';
+import { createESLintRule } from '../utils/create-eslint-rule';
 import { getDomElements } from '../utils/get-dom-elements';
 import { toPattern } from '../utils/to-pattern';
