+++ packages/eslint-plugin-template/src/utils/create-eslint-rule.ts
@@ -1,15 +1,12 @@
-import type {
-  ParseSourceSpan,
-  TmplAstElement,
-} from '@angular-eslint/bundled-angular-compiler';
-import type { TSESLint, TSESTree } from '@typescript-eslint/utils';
 import { ESLintUtils } from '@typescript-eslint/utils';
 
 /**
  * We need to patch the RuleCreator in order to preserve the defaultOptions
  * to use as part of documentation generation.
  */
-const patchedRuleCreator: typeof ESLintUtils.RuleCreator = (urlCreator) => {
+export const patchedRuleCreator: typeof ESLintUtils.RuleCreator = (
+  urlCreator,
+) => {
   return function createRule({ name, meta, defaultOptions, create }) {
     return {
       meta: Object.assign(Object.assign({}, meta), {
@@ -35,43 +32,3 @@ export const createESLintRule = patchedRuleCreator(
   (ruleName) =>
     `https://github.com/angular-eslint/angular-eslint/blob/main/packages/eslint-plugin-template/docs/rules/${ruleName}.md`,
 );
-
-interface ParserServices {
-  convertNodeSourceSpanToLoc: (
-    sourceSpan: ParseSourceSpan,
-  ) => TSESTree.SourceLocation;
-  convertElementSourceSpanToLoc: (
-    context: Readonly<TSESLint.RuleContext<string, readonly unknown[]>>,
-    node: TmplAstElement,
-  ) => TSESTree.SourceLocation;
-}
-
-export function getTemplateParserServices(
-  context: Readonly<TSESLint.RuleContext<string, readonly unknown[]>>,
-): ParserServices {
-  ensureTemplateParser(context);
-  return context.parserServices as unknown as ParserServices;
-}
-
-/**
- * Utility for rule authors to ensure that their rule is correctly being used with @angular-eslint/template-parser
- * If @angular-eslint/template-parser is not the configured parser when the function is invoked it will throw
- */
-export function ensureTemplateParser(
-  context: Readonly<TSESLint.RuleContext<string, readonly unknown[]>>,
-): void {
-  if (
-    !(context.parserServices as unknown as ParserServices)
-      ?.convertNodeSourceSpanToLoc ||
-    !(context.parserServices as unknown as ParserServices)
-      ?.convertElementSourceSpanToLoc
-  ) {
-    /**
-     * The user needs to have configured "parser" in their eslint config and set it
-     * to @angular-eslint/template-parser
-     */
-    throw new Error(
-      "You have used a rule which requires '@angular-eslint/template-parser' to be used as the 'parser' in your ESLint config.",
-    );
-  }
-}
